requirements.txt
google-genai>=1.0.0
google-adk>=0.3.0
nemoguardrails>=0.10.0
python-dotenv>=1.0.0
openai>=1.0.0
litellm!=1.82.7,!=1.82.8
